First commit of variant reconstruction code
Reads in files containing snps with pos, base, number of reads with
snp, and read depth (snps_based_on_basefreq.out) and read id with snp
that is in that read (hapread.out). Identifies snps that have low
frequency, are in a low read depth region, and are not found to be
connected to other snps. Removes this snps from the list and outputs a
pruned list of snps with the adjacency matrix (called link matrix) from
these pruned snps. Also does a BFS search on the adjacency matrix to
find connected components. Future changes in code: change from BFS
search to clique search (fully connected components).